# API/openAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Remote Ops API",
    "description": "API cho phép thực hiện các thao tác từ xa - file, quét dữ liệu và chạy lệnh",
    "version": "1.0.0",
    "contact": {
      "name": "Remote Ops API System"
    }
  },
  "servers": [
    {
      "url": "https://climbing-lioness-factual.ngrok-free.app",
      "description": "Remote Server"
    }
  ],
...
  "paths": {
...
    "/files/list": {
      "get": {
        "operationId": "list_directory",
        "tags": [
          "FileOps"
        ],
        "summary": "Liệt kê nội dung thư mục",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "Đường dẫn tới thư mục cần liệt kê",
            "schema": {
              "type": "string",
              "example": "C:/Users"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách các file và thư mục",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Directory listed successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "example": "example.txt"
                          },
                          "path": {
                            "type": "string",
                            "example": "C:/Users/example.txt"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "file",
                              "directory"
                            ]
                          },
                          "size": {
                            "type": "integer",
                            "nullable": true,
                            "example": 1024
                          },
                          "modified": {
                            "type": "number",
                            "format": "float",
                            "example": 1653422485.234
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy thư mục"
          }
        }
      }
    },
    "/files/create/file": {
      "post": {
        "operationId": "create_file",
        "tags": [
          "FileOps"
        ],
        "summary": "Tạo file mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "path"
                ],
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "Đường dẫn tới file cần tạo",
                    "example": "C:/path/to/file.txt"
                  },
                  "content": {
                    "type": "string",
                    "description": "Nội dung file (tùy chọn)",
                    "example": "This is content of the file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File đã được tạo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "File created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "path": {
                          "type": "string",
                          "example": "C:/path/to/file.txt"
                        },
                        "size": {
                          "type": "integer",
                          "example": 24
                        },
                        "created": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/files/create/directory": {
      "post": {
        "operationId": "create_directory",
        "tags": [
          "FileOps"
        ],
        "summary": "Tạo thư mục mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "path"
                ],
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "Đường dẫn thư mục cần tạo",
                    "example": "C:/path/to/new_folder"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thư mục đã được tạo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Directory created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "path": {
                          "type": "string",
                          "example": "C:/path/to/new_folder"
                        },
                        "created": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/files/read": {
      "get": {
        "operationId": "read_file",
        "tags": [
          "FileOps"
        ],
        "summary": "Đọc nội dung file",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "Đường dẫn đến file cần đọc",
            "schema": {
              "type": "string",
              "example": "C:/path/to/file.txt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nội dung file",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "File read successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "path": {
                          "type": "string",
                          "example": "C:/path/to/file.txt"
                        },
                        "content": {
                          "type": "string",
                          "example": "Content of the file"
                        },
                        "size": {
                          "type": "integer",
                          "example": 24
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy file"
          }
        }
      }
    },
    "/files/delete": {
      "delete": {
        "operationId": "delete_item",
        "tags": [
          "FileOps"
        ],
        "summary": "Xóa file hoặc thư mục",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "Đường dẫn đến file hoặc thư mục cần xóa",
            "schema": {
              "type": "string",
              "example": "C:/path/to/file.txt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Đã xóa thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "file deleted successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "path": {
                          "type": "string",
                          "example": "C:/path/to/file.txt"
                        },
                        "deleted": {
                          "type": "boolean",
                          "example": true
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "file",
                            "directory"
                          ],
                          "example": "file"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy file hoặc thư mục"
          }
        }
      }
    },
    "/files/move": {
      "post": {
        "operationId": "move_item",
        "tags": [
          "FileOps"
        ],
        "summary": "Di chuyển hoặc đổi tên file/thư mục",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "source_path",
                  "destination_path"
                ],
                "properties": {
                  "source_path": {
                    "type": "string",
                    "description": "Đường dẫn nguồn",
                    "example": "C:/path/source.txt"
                  },
                  "destination_path": {
                    "type": "string",
                    "description": "Đường dẫn đích",
                    "example": "C:/new_path/dest.txt"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã di chuyển thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "file moved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "source": {
                          "type": "string",
                          "example": "C:/path/source.txt"
                        },
                        "destination": {
                          "type": "string",
                          "example": "C:/new_path/dest.txt"
                        },
                        "moved": {
                          "type": "boolean",
                          "example": true
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "file",
                            "directory"
                          ],
                          "example": "file"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy file hoặc thư mục nguồn"
          }
        }
      }
    },
    "/files/copy": {
      "post": {
        "operationId": "copy_item",
        "tags": [
          "FileOps"
        ],
        "summary": "Sao chép file/thư mục",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "source_path",
                  "destination_path"
                ],
                "properties": {
                  "source_path": {
                    "type": "string",
                    "description": "Đường dẫn nguồn",
                    "example": "C:/path/source.txt"
                  },
                  "destination_path": {
                    "type": "string",
                    "description": "Đường dẫn đích",
                    "example": "C:/new_path/dest.txt"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã sao chép thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "file copied successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "source": {
                          "type": "string",
                          "example": "C:/path/source.txt"
                        },
                        "destination": {
                          "type": "string",
                          "example": "C:/new_path/dest.txt"
                        },
                        "copied": {
                          "type": "boolean",
                          "example": true
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "file",
                            "directory"
                          ],
                          "example": "file"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy file hoặc thư mục nguồn"
          }
        }
      }
    },
    "/files/info": {
      "get": {
        "operationId": "get_file_info",
        "tags": [
          "FileOps"
        ],
        "summary": "Lấy thông tin chi tiết về file/thư mục",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "Đường dẫn đến file hoặc thư mục",
            "schema": {
              "type": "string",
              "example": "C:/path/to/file.txt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin chi tiết",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "File info retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "path": {
                          "type": "string",
                          "example": "C:/path/to/file.txt"
                        },
                        "name": {
                          "type": "string",
                          "example": "file.txt"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "file",
                            "directory"
                          ],
                          "example": "file"
                        },
                        "size": {
                          "type": "integer",
                          "nullable": true,
                          "example": 1024
                        },
                        "created": {
                          "type": "number",
                          "format": "float",
                          "example": 1653422485.234
                        },
                        "modified": {
                          "type": "number",
                          "format": "float",
                          "example": 1653422485.234
                        },
                        "accessed": {
                          "type": "number",
                          "format": "float",
                          "example": 1653422485.234
                        },
                        "permissions": {
                          "type": "integer",
                          "example": 33206
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy file hoặc thư mục"
          }
        }
      }
    },
...
  },
...
}